Read stream metadata from CSV file and use it to describe the streams in the Waveform messages.

+++ emap-interchange/src/main/java/uk/ac/ucl/rits/inform/interchange/visit_observations/WaveformMessage.java
@@ -55,7 +55,7 @@ public class WaveformMessage extends EmapOperationMessage {
     /**
      * Sampling rate in Hz.
      */
-    private long samplingRate;
+    private int samplingRate;
 
     /**
      * Numeric value.
